feat: completed projects-main-section adaptive
