+++ src/app-data.json
@@ -40,6 +40,7 @@
         "id": "mini-projects",
         "name": "Мини-проекты",
         "desk": "Мини-проекты, созданные в данном приложении с резюме: ToDo List, Weather App",
+        "adaptiveDesk": "Мини-проекты, созданные в данном приложении с резюме: ToDo List, Weather App",
         "git": "https://github.com/Vagoalex/vago-website-resume",
         "title": "Нажми, чтобы перейти на git репозиторий всего приложения :)",
         "seeLive": "/project-page",
@@ -49,7 +50,8 @@
         {
           "id": "marvel-app",
           "name": "Marvel App",
-          "desk": "Это проект - Marvel App, с него я начинал свое обучение реакту. Тут нет Redux или MobX, тут state management работает на обычных хуках. Реализовано 2 страницы - Characters и Comics. Через роутер есть переход по ним. Запросы персонажей и комиксов идут через marvel api. Есть форма для поиска отдельного персонажа, с валидацией через Formik. Также есть отдельные странички для одного комикса или персонажа. Есть прикольные штуки с загрузками. Начал адаптив, добрался до 1000px и забил, когда нибудь сделаю.",
+          "desk": "С него я начинал свое обучение реакту. Тут нет Redux или MobX, тут state management работает на обычных хуках. Реализовано 2 страницы - Characters и Comics. Через роутер есть переход по ним. Запросы персонажей и комиксов идут через marvel api. Есть форма для поиска отдельного персонажа, с валидацией через Formik. Также есть отдельные странички для одного комикса или персонажа. Есть прикольные штуки с загрузками. Начал адаптив, добрался до 1000px и забил, когда нибудь сделаю.",
+          "adaptiveDesk": "На нем изучал чистый React без state managements. Есть страницы Characters и Comics. Также есть отдельные странички для персонажа из формы и комикса из списка комиксов. В проекте используется marvel api для запросов.",
           "git": "https://github.com/Vagoalex/marvelApp__React",
           "title": "Нажми, чтобы перейти на git репозиторий :)",
           "seeLive": "https://vagoalex.github.io/marvelApp__React/#/",
@@ -58,7 +60,8 @@
         {
           "id": "hero-panel",
           "name": "Hero Admin Panel",
-          "desk": "Это проект - Hero Admin Panel, который работает только на json-server и localhost. Там есть мини база данных, куда приходят данные и удаляются оттуда. Если коротко, то это небольшая админ панель, где мы можем создать героя, выбрать ему определенные атрибуты, и добавить его в список через форму (обязательно с валидацией), также есть фильтры по данным атрибутам. Использовался чистый Redux и обновленная версия с использованием Toolkit.",
+          "desk": "Это проект, который работает только на json-server и localhost. Там есть мини база данных, куда приходят данные и удаляются оттуда. Если коротко, то это небольшая админ панель, где мы можем создать героя, выбрать ему определенные атрибуты, и добавить его в список через форму (обязательно с валидацией), также есть фильтры по данным атрибутам. Использовался чистый Redux и обновленная версия с использованием Toolkit.",
+          "adaptiveDesk": "Этот проект работает только на json-server и localhost, с базой данных. Админ панель, где можно создать героя с атрибутами. Можно фильтровать или удалять героев. React, Redux + Toolkit",
           "git": "https://github.com/Vagoalex/hero_admin_panel",
           "title": "Нажми, чтобы перейти на git репозиторий :)",
           "seeLive": null,
