+++ src/components/pages/main-page/projects/Projects.scss
@@ -1,14 +1,17 @@
+@import '../../../../style/variables.scss';
+
 .Projects-section {
   position: relative;
   text-align: center;
   background: #191919;
   padding: 150px 0px;
-  transition: 1s;
+  transition: all 1s ease;
 }
 
 .Projects-title-container {
   position: relative;
   display: inline-block;
+  transition: all 1s ease;
 
   .Projects-title {
     display: flex;
@@ -35,6 +38,19 @@
           1px 1px 0 rgb(144, 15, 230), 1px -1px 0 rgb(144, 15, 230),
           -1px 1px 0 rgb(144, 15, 230), -1px -1px 0 rgb(144, 15, 230);
       }
+
+      @media (max-width: $mediaSmallDesktops) {
+        font-size: 80px;
+      }
+      @media (max-width: $mediaSmallTablets) {
+        font-size: 50px;
+      }
+      @media (max-width: $mediaPhones) {
+        font-size: 40px;
+      }
+      @media (max-width: $mediaSmallPhones) {
+        font-size: 30px;
+      }
     }
     .projects-item-hidden {
       text-shadow: none;
@@ -51,6 +67,15 @@
     width: 100%;
     height: 3px;
     background-color: rgb(124, 49, 173);
+    @media (max-width: $mediaSmallTablets) {
+      top: -30px;
+    }
+    @media (max-width: $mediaPhones) {
+      top: -20px;
+    }
+    @media (max-width: $mediaSmallPhones) {
+      top: -15px;
+    }
   }
   &:after {
     content: '';
@@ -61,6 +86,15 @@
     width: 100%;
     height: 3px;
     background-color: rgb(124, 49, 173);
+    @media (max-width: $mediaSmallTablets) {
+      bottom: -35px;
+    }
+    @media (max-width: $mediaPhones) {
+      bottom: -25px;
+    }
+    @media (max-width: $mediaSmallPhones) {
+      bottom: -18px;
+    }
   }
 }
 
@@ -69,4 +103,12 @@
   grid-template-columns: repeat(2, 1fr);
   grid-gap: 100px;
   margin-top: 100px;
+
+  transition: all 1s ease;
+
+  @media (max-width: $mediaSmallTablets) {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
 }
