+++ src/components/pages/main-page/projects/project-card/ProjectCard.jsx
@@ -1,25 +1,22 @@
 import { useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
-
-import miniProjectsImg from '../../../../../assets/images/main-page/mini-projects.jpg';
-import heroAdminImg from '../../../../../assets/images/main-page/project-hero-admin-panel.jpg';
-import marvellAppImg from '../../../../../assets/images/main-page/project-marvel-app.jpg';
-
-import miniProjectsImgAdaptive from '../../../../../assets/images/main-page/mini-projects-adaptive.jpg';
-import heroAdminImgAdaptive from '../../../../../assets/images/main-page/project-hero-admin-panel-adaptive.jpg';
-import marvellAppImgAdaptive from '../../../../../assets/images/main-page/project-marvel-app-adaptive.jpg';
+import ImageProjectCard from './ImageProjectCard';
+import LinkLiveCard from './LinkLiveCard';
 
 import './ProjectCard.scss';
 
 const ProjectCard = ({ data }) => {
-  const { desk, git, id, name, seeLive, seeLiveTitle, title } = data;
+  const { screenWidth } = useSelector((state) => state.nav);
+  const { desk, adaptiveDesk, git, id, name, seeLive, seeLiveTitle, title } =
+    data;
 
   return (
     <div className='Project-card'>
       <ImageProjectCard id={id} />
       <div className='Project-content'>
         <h2 className='Project-title'>{name}</h2>
-        <p className='Project-info'>{desk}</p>
+        <p className='Project-info'>
+          {screenWidth < 991 ? adaptiveDesk : desk}
+        </p>
         <div className='Project-btn-grp'>
           <a
             href={git}
@@ -35,80 +32,6 @@ const ProjectCard = ({ data }) => {
       </div>
     </div>
   );
-};
-
-const ImageProjectCard = ({ id }) => {
-  const { activeScreenBurgerMenu } = useSelector((state) => state.nav);
-
-  switch (id) {
-    case 'mini-projects':
-      return (
-        <img
-          src={
-            activeScreenBurgerMenu ? miniProjectsImgAdaptive : miniProjectsImg
-          }
-          className='Project-img'
-          alt='hero-admin-img'
-        />
-      );
-    case 'marvel-app':
-      return (
-        <img
-          src={activeScreenBurgerMenu ? marvellAppImgAdaptive : marvellAppImg}
-          className='Project-img'
-          alt='hero-admin-img'
-        />
-      );
-    case 'hero-panel':
-      return (
-        <img
-          src={activeScreenBurgerMenu ? heroAdminImgAdaptive : heroAdminImg}
-          className='Project-img'
-          alt='hero-admin-img'
-        />
-      );
-
-    default:
-      throw new Error('Invalid id');
-  }
-};
-
-const LinkLiveCard = ({ id, link, liveTitle }) => {
-  switch (id) {
-    case 'mini-projects':
-      return (
-        <Link to={link} className='Project-btn live' title={liveTitle}>
-          see live
-        </Link>
-      );
-    case 'marvel-app':
-      return (
-        <a
-          href={link}
-          className='Project-btn live'
-          target='_blank'
-          rel='noreferrer'
-          title={liveTitle}
-        >
-          see live
-        </a>
-      );
-    case 'hero-panel':
-      return (
-        <a
-          href={link}
-          className='Project-btn live Project-btn--disabled'
-          target='_blank'
-          rel='noreferrer'
-          title={liveTitle}
-        >
-          see live
-        </a>
-      );
-
-    default:
-      throw new Error('Invalid id');
-  }
 };
 
 export default ProjectCard;
