+++ src/components/pages/main-page/projects/project-card/ProjectCard.scss
@@ -1,6 +1,25 @@
+@import '../../../../../style/variables.scss';
+
 .Project-card {
   height: 400px;
   position: relative;
+
+  transition: all 1s ease;
+
+  @media (max-width: $mediaSmallDesktops) {
+    height: 275px;
+  }
+  @media (max-width: $mediaSmallTablets) {
+    height: 325px;
+  }
+  @media (max-width: $mediaPhones) {
+    height: 310px;
+    width: 310px;
+  }
+  @media (max-width: $mediaSmallPhones) {
+    height: 250px;
+    width: 250px;
+  }
 }
 .Project-img {
   width: 100%;
@@ -21,23 +40,58 @@
   transition: 0.5s;
 
   opacity: 0;
+
+  @media (max-width: $mediaSmallDesktops) {
+    height: 275px;
+  }
+  @media (max-width: $mediaSmallTablets) {
+    height: 325px;
+  }
+  @media (max-width: $mediaPhones) {
+    height: 310px;
+  }
+  @media (max-width: $mediaSmallPhones) {
+    height: 250px;
+  }
 }
 .Project-title {
   font-size: 40px;
   text-transform: capitalize;
   text-align: center;
   font-weight: 300;
+
+  @media (max-width: $mediaSmallDesktops) {
+    font-size: 28px;
+  }
 }
 .Project-info {
   margin: 40px;
   font-size: 16px;
   line-height: 20px;
   text-align: justify;
+
+  @media (max-width: $mediaSmallDesktops) {
+    margin: 10px;
+  }
+  @media (max-width: $mediaSmallTablets) {
+    margin: 35px 10px;
+  }
+  @media (max-width: $mediaSmallPhones) {
+    font-size: 13px;
+    margin: 15px 10px;
+  }
 }
 .Project-btn-grp {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   grid-gap: 20px;
+
+  @media (max-width: $mediaSmallDesktops) {
+    position: absolute;
+    left: 0;
+    bottom: 0;
+    right: 0;
+  }
 }
 .Project-btn {
   display: flex;
